requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==19.3.0
Automat==0.8.0
azure-core==1.30.1
azure-storage-blob==12.19.1
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.3
bitarray==2.9.2
blinker==1.4
bottle==0.12.15
Bottleneck==1.3.8
build==1.2.1
cachetools==5.3.3
casadi==3.6.7
certifi==2019.11.28
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
cloud-init==23.1.2
cloudpickle==3.0.0
colorama==0.4.4
coloredlogs==15.0.1
colorlog==4.7.2
comm==0.2.2
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.1.1
cryptography==2.8
cupshelpers==1.0
cupy-cuda12x==12.3.0
cycler==0.10.0
dask==2023.5.0
datasketch==1.6.4
dbus-python==1.2.16
debugpy==1.8.1
decorator==5.1.1
defer==1.0.6
Deprecated==1.2.14
distro==1.4.0
distro-info==0.23ubuntu1
dnspython==2.6.1
docker==4.1.0
email-validator==2.1.1
entrypoints==0.3
environs==9.5.0
exceptiongroup==1.2.1
executing==2.0.1
faiss-cpu==1.8.0
fastapi==0.111.0
fastapi-cli==0.0.2
fastrlock==0.8.2
filelock==3.13.1
flatbuffers==24.3.7
fonttools==4.53.0
fsspec==2023.12.2
future==0.18.2
gast==0.4.0
Glances==3.1.3
google-auth==2.28.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.63.0
GPUtil==1.4.0
graphlib-backport==1.1.0
greenlet==0.4.15
grpcio==1.60.0
grpcio-tools==1.63.0
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httplib2==0.14.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
humanfriendly==10.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==7.0.2
importlib-resources==6.4.0
incremental==16.10.1
influxdb==5.2.0
iotop==0.6
ipykernel==6.29.3
ipython==8.12.3
isodate==0.6.1
jedi==0.19.1
jinja2==3.1.4
joblib==1.3.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==8.6.1
jupyter-core==5.7.2
keras==2.13.1
keyring==18.0.1
kiwisolver==1.0.1
kubernetes==29.0.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==18.1.1
line-profiler==4.1.3
locket==1.0.0
macaroonbakery==1.3.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.7.5
matplotlib-inline==0.1.6
mdurl==0.1.2
milvus==2.3.5
minio==7.2.7
mmh3==4.1.0
monotonic==1.6
more-itertools==4.2.0
mpmath==1.3.0
msgpack==0.6.2
nest-asyncio==1.6.0
netifaces==0.10.4
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
objsize==0.7.0
olefile==0.46
onnxruntime==1.16.3
opencv-python==4.10.0.82
opencv-python-headless==4.10.0.82
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
opt-einsum==3.3.0
orjson==3.10.3
overrides==7.7.0
packaging==24.0
pandas==1.1.4
parso==0.8.3
partd==1.4.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pip==20.0.2
platformdirs==4.2.0
plotly==5.20.0
ply==3.11
posthog==3.5.0
prompt-toolkit==3.0.43
protobuf==5.26.1
psutil==5.5.1
PuLP==2.9.0
pure-eval==0.2.2
pyarrow==16.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.6.1
pycups==1.9.73
pydantic==2.7.1
pydantic-core==2.18.2
pygments==2.17.2
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymemcache==4.0.0
pymilvus==2.4.1
Pympler==1.0.1
PyNaCl==1.3.0
pynvim==0.4.1
pynvml==11.4.1
pyOpenSSL==19.0.0
pyparsing==2.4.6
PyPika==0.48.9
pyproject-hooks==1.1.0
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
pysmi==0.3.2
pysnmp==4.4.6
pysqlite3-binary==0.5.2.post3
pystache==0.5.4
python-apt==2.0.0+ubuntu0.20.4.6
python-box==7.1.1
python-dateutil==2.8.1
python-debian==0.1.36ubuntu1
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2020.4
PyYAML==6.0.1
pyzmq==25.1.2
recbole==1.2.0
requests==2.31.0
requests-oauthlib==1.4.0
requests-unixsocket==0.2.0
rich==13.7.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==2.3.1
service-identity==18.1.0
setuptools==69.5.1
shellingham==1.5.4
simplejson==3.16.0
six==1.14.0
sniffio==1.3.1
sos==4.3
ssh-import-id==5.10
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
systemd-python==234
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
texttable==1.7.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tokenizers==0.19.1
tomli==2.0.1
toolz==0.12.1
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.2
triton==2.1.0
Twisted==18.9.0
typer==0.12.3
typing-extensions==4.11.0
tzdata==2024.1
ubuntu-advantage-tools==27.5
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==5.9.0
unattended-upgrades==0.1
urllib3==1.25.8
uvicorn==0.29.0
uvloop==0.19.0
wadllib==1.3.3
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==0.53.0
websockets==12.0
werkzeug==3.0.1
wheel==0.34.2
wrapt==1.16.0
xkit==0.0.0
zipp==3.18.1
zope.interface==4.7.1
